requirements_test.txt
#pytest-homeassistant-custom-component==0.3.2
pytest-homeassistant-custom-component @ git+https://github.com/MatthewFlamm/pytest-homeassistant-custom-component@0.13.87 # HA core 2024.1.0